+++ src/index.js
@@ -20,7 +20,7 @@ const plugin = {
 	meta: {
 		name: "@eslint/css",
 		namespace: "css",
-		version: "1.3.0", // x-release-please-version
+		version: "1.4.0", // x-release-please-version
 	},
 	languages: {
 		css: new CSSLanguage(),