+++ src/rules/no-invalid-properties.js
@@ -86,8 +86,7 @@ function getVarFallbackList(value) {
 // Rule Definition
 //-----------------------------------------------------------------------------
 
-/** @type {NoInvalidPropertiesRuleDefinition} */
-export default {
+export default /** @satisfies {NoInvalidPropertiesRuleDefinition} */ ({
 	meta: {
 		type: "problem",
 		languages: ["css/css"],
@@ -492,4 +491,4 @@ export default {
 			},
 		};
 	},
-};
+});