+++ src/rules/selector-complexity.js
@@ -152,8 +152,7 @@ function getDisallowedCombinatorsLocation(
 // Rule Definition
 //-----------------------------------------------------------------------------
 
-/** @type {SelectorComplexityRuleDefinition} */
-export default {
+export default /** @satisfies {SelectorComplexityRuleDefinition} */ ({
 	meta: {
 		type: "problem",
 		languages: ["css/css"],
@@ -499,4 +498,4 @@ export default {
 			},
 		};
 	},
-};
+});