+++ src/rules/use-baseline.js
@@ -483,8 +483,7 @@ class BaselineAvailability {
 // Rule Definition
 //-----------------------------------------------------------------------------
 
-/** @type {UseBaselineRuleDefinition} */
-export default {
+export default /** @satisfies {UseBaselineRuleDefinition} */ ({
 	meta: {
 		type: "problem",
 		languages: ["css/css"],
@@ -1050,4 +1049,4 @@ export default {
 			},
 		};
 	},
-};
+});