+++ tests/rules/relative-font-units.test.js
@@ -50,6 +50,8 @@ ruleTester.run("relative-font-units", rule, {
 		"a { font: revert-layer Arial, sans-serif; }",
 		"a { font-size: unset; }",
 		"a { font: unset Arial, sans-serif; }",
+		"a { font: 1rem/120% Arial, sans-serif; }",
+		"a { font: italic smaller Arial, sans-serif; }",
 		{
 			code: "a { font-size: 1em; }",
 			options: [
@@ -859,6 +861,24 @@ ruleTester.run("relative-font-units", rule, {
 		{
 			code: dedent`
                 a {
+                    font: 1rem/120% Arial, sans-serif;
+                }
+            `,
+			options: [
+				{
+					allowUnits: ["em"],
+				},
+			],
+			errors: [
+				{
+					messageId: "allowedFontUnits",
+					data: { allowedFontUnits: "em" },
+				},
+			],
+		},
+		{
+			code: dedent`
+                a {
                     font: 2em Arial, sans-serif;
                 }
             `,
@@ -1233,6 +1253,15 @@ ruleTester.run("relative-font-units", rule, {
 		},
 		{
 			code: "a { font: Small Arial, sans-serif; }",
+			errors: [
+				{
+					messageId: "allowedFontUnits",
+					data: { allowedFontUnits: "rem" },
+				},
+			],
+		},
+		{
+			code: "a { font: italic small Arial, sans-serif; }",
 			errors: [
 				{
 					messageId: "allowedFontUnits",